+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 All notable changes to this project are documented in this file.
 
-## X.X.X
+## 0.5.0
 
 - Exception handling during the Auto Reload
 - Last column sorting honored on reload
+- Kubectl context switch dropdown menu
+- Minor improvements
 
 ## 0.4.1